+++ resources/views/auth/setup.blade.php
@@ -18,7 +18,8 @@
 
         <label class="auth-field">
             <span>{{ __('Password') }}</span>
-            <input type="password" name="password" required autocomplete="new-password">
+            <input type="password" name="password" required autocomplete="new-password"
+                maxlength="{{ config('auth.password_max_length') }}">
         </label>
         @error('password')
             <div class="auth-error">{{ $message }}</div>
@@ -26,7 +27,8 @@
 
         <label class="auth-field">
             <span>{{ __('Confirm password') }}</span>
-            <input type="password" name="password_confirmation" required autocomplete="new-password">
+            <input type="password" name="password_confirmation" required autocomplete="new-password"
+                maxlength="{{ config('auth.password_max_length') }}">
         </label>
 
         <p class="auth-hint">{{ __('Use at least 8 characters.') }}</p>